# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Tasks API",
        "description": "API REST para gestión de tareas — Challenge IT Rock",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Lista tareas del usuario autenticado",
                "description": "Muestra la lista paginada de tareas del usuario.",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtra por prioridad (low, medium, high)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "medium"
                        }
                    },
                    {
                        "name": "completed",
                        "in": "query",
                        "description": "Filtra por estado de completado (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedTasks"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crea una tarea",
                "description": "Crea una nueva tarea para el usuario autenticado.",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{task}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtiene una tarea",
                "description": "Muestra el detalle de una tarea.",
                "operationId": "bd73ed9e6d55cbc8a547d4aac67e2ea5",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "UUID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualiza una tarea",
                "description": "Actualiza una tarea existente.",
                "operationId": "a6b00b06431eae3ece4a210a52501fbd",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "UUID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Elimina una tarea",
                "description": "Elimina una tarea del usuario.",
                "operationId": "5626a232c6ca6481a76585b267774c10",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "UUID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{task}/complete": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Marca una tarea como completada",
                "description": "Marca una tarea como completada.",
                "operationId": "53d37a01f55799abf147ec0995beb244",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{task}/incomplete": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Marca una tarea como incompleta",
                "description": "Marca una tarea como incompleta.",
                "operationId": "a9d931e69d5acc844d5919781bf7ccc4",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/populate": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Importa tareas externas",
                "description": "Sincroniza tareas desde la API externa jsonplaceholder.",
                "operationId": "269868c0666a045231a34d9c9e211686",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Clave de seguridad registrada en API_POPULATE_KEY",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sincronización exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "inserted": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "API key inválida"
                    },
                    "502": {
                        "description": "Error al consumir el servicio externo"
                    },
                    "500": {
                        "description": "Fallo interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "description": "Entidad Task",
                "required": [
                    "id",
                    "title",
                    "is_completed",
                    "priority",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "6f5cbd4d-9a9e-4d3f-a3e4-33f420c65d7b"
                    },
                    "title": {
                        "type": "string",
                        "example": "Preparar presentación"
                    },
                    "description": {
                        "type": "string",
                        "example": "Actualizar slides para la demo",
                        "nullable": true
                    },
                    "is_completed": {
                        "type": "boolean",
                        "example": false
                    },
                    "priority": {
                        "type": "string",
                        "example": "medium"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TaskRequest": {
                "description": "Payload para crear o actualizar tareas",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Comprar insumos"
                    },
                    "description": {
                        "type": "string",
                        "example": "Comprar café y snacks"
                    },
                    "is_completed": {
                        "type": "boolean",
                        "example": false
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "high"
                    }
                },
                "type": "object"
            },
            "PaginatedTasks": {
                "description": "Respuesta paginada estándar de Laravel",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8080/api/tasks?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8080/api/tasks?page=5"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 5
                            },
                            "path": {
                                "type": "string",
                                "example": "http://localhost:8080/api/tasks"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "to": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 65
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}